Removed duplicate conversation.rs and main.rs
